feat(tally): Change username to email
Only emails are accepted as usernames since they are needed later to send emails to the users. On register a new user the name is checked to be an email address.
The backend also checks if the email is valid and saves them in the renamed database column 'email'.
The user 'admin' is not created anymore and instead admins are configured in the application properties.

Issue: #28