feat(tally): Send registration email on user register
When a new user registers, send a registration email. The email contains a link that finishes the registration. A user can only log in after the registration process is finished by clicking on the link in the email. The link has the email and a secret, encoded in base64 and URL safe appended.

Issue: #32